sync(bsw): stamp BswSchedulerNamePrefix spec verified R23-11
Adds # Spec verified: R23-11 marker for BswSchedulerNamePrefix (Table 5.20, p.86)
after Step 9a automated checks + Step 9b rule-compliance gate passed with no
deviations. Moves the class from the standalone BswSchedulerNamePrefix.py leaf
module into BswBehavior/__init__.py (the correct package file, consistent with
its sibling BswBehavior classes); updates imports and import-discovery test.
